Convert Primitive::Int and SqInt to use i128
Let's use 128-bit integers for our generic integer types so that they
can hold both u64::MAX and i64::MIN.

For #53: Add more primitive integer types.

+++ src/ast.rs
@@ -99,7 +99,7 @@ pub enum Filter {
 #[derive(Clone, Debug, PartialEq)]
 pub struct IntLiteral {
     pub span: SourceSpan,
-    pub value: i64,
+    pub value: i128,
 }
 
 /// AST node representing a `slice` grammar item in a `filter_pack`: a Python style slice.
